Feature/oauth2

+++ src/main/java/com/gary/demo/crud/service/DemoCrudServiceImpl.java
@@ -4,9 +4,9 @@ import com.gary.demo.crud.api.exception.DataNotFoundException;
 import com.gary.demo.crud.api.exception.ValidationException;
 import com.gary.demo.crud.dao.PersonDao;
 import com.gary.demo.crud.model.Person;
+import com.gary.demo.crud.model.PersonResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
@@ -20,13 +20,13 @@ public class DemoCrudServiceImpl implements DemoCrudService{
     PersonDao personDao;
 
     @Override
-    public ResponseEntity<Person> getPerson(String id) {
+    public PersonResponse getPerson(String id) {
         if(StringUtils.isEmpty(id)){
             throw new ValidationException("Missing ID");
         }
         Optional<com.gary.demo.crud.entity.Person> optionalPersonEntity = personDao.findById(Long.parseLong(id));
         if(optionalPersonEntity.isPresent()){
-            return ResponseEntity.ok(convertToPersonDto(optionalPersonEntity.get()));
+            return new PersonResponse(HttpStatus.OK, "Successful", convertToPersonDto(optionalPersonEntity.get()));
 
         }
         else{
@@ -36,19 +36,18 @@ public class DemoCrudServiceImpl implements DemoCrudService{
     }
 
     @Override
-    public ResponseEntity<Person> createPerson(Person person) {
+    public PersonResponse createPerson(Person person) {
         if(StringUtils.isEmpty(person.getFirstName()) || StringUtils.isEmpty(person.getLastName())){
             throw new ValidationException("Missing First Name or Last Name");
         }
 
         com.gary.demo.crud.entity.Person personEntity = personDao.save(convertToPersonEntity(person));
-        return ResponseEntity.status(HttpStatus.CREATED).body(convertToPersonDto(personEntity));
-
+        return new PersonResponse(HttpStatus.CREATED, "Created", convertToPersonDto(personEntity));
     }
 
     @Override
     @Transactional
-    public ResponseEntity<Person> updatePerson(Person person) {
+    public PersonResponse updatePerson(Person person) {
         if(StringUtils.isEmpty(person.getId())){
             throw new ValidationException("Missing ID");
         }
@@ -56,18 +55,31 @@ public class DemoCrudServiceImpl implements DemoCrudService{
             throw new ValidationException("Must at least have either First Name or Last Name");
         }
 
-        com.gary.demo.crud.entity.Person personEntity = personDao.save(convertToPersonEntity(person));
-        return ResponseEntity.ok(convertToPersonDto(personEntity));
+        Optional<com.gary.demo.crud.entity.Person> optionalPersonEntity = personDao.findById(Long.parseLong(person.getId()));
+        if(!optionalPersonEntity.isPresent()){
+            //throw 404 not found
+            throw new DataNotFoundException("Person with id " + person.getId() + " does not exist");
+        }
+        com.gary.demo.crud.entity.Person personEntity = optionalPersonEntity.get();
+        personEntity.setFirstName(person.getFirstName());
+        personEntity.setLastName(person.getLastName());
+        personEntity = personDao.save(personEntity);
+        return new PersonResponse(HttpStatus.OK, "Updated", convertToPersonDto(personEntity));
     }
 
     @Override
     @Transactional
-    public ResponseEntity<Person> deletePerson(String id) {
+    public PersonResponse deletePerson(String id) {
         if(StringUtils.isEmpty(id)){
             throw new ValidationException("Missing ID");
         }
+        Optional<com.gary.demo.crud.entity.Person> optionalPersonEntity = personDao.findById(Long.parseLong(id));
+        if(!optionalPersonEntity.isPresent()){
+            //throw 404 not found
+            throw new DataNotFoundException("Person with id " + id + " does not exist");
+        }
         personDao.deleteById(Long.parseLong(id));
-        return ResponseEntity.ok(null);
+        return new PersonResponse(HttpStatus.OK, "Deleted");
     }
 
     private com.gary.demo.crud.model.Person convertToPersonDto(com.gary.demo.crud.entity.Person personEntity){
